+++ Kanikama/Assets/Kanikama/Scripts/Udon/KanikamaColorCollector.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d4f0981d8afc1ff418a14fe2d7ed66eb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
