+++ Kanikama/Assets/Kanikama/Scripts/Udon/KanikamaColorCollector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 28addd43541785942a5553970ab321b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
